+++ apps/server/src/binaries/binaries.module.ts
@@ -3,12 +3,13 @@ import { BinariesController } from './binaries.controller';
 import { BinariesService } from './binaries.service';
 import { BinariesInstallerService } from './installer.service';
 import { BinaryResourcesService } from './binary-resources.service';
+import { OfflinePackageService } from './offline-package.service';
 import { SystemModule } from '../system/system.module';
 
 @Module({
   imports: [SystemModule],
   controllers: [BinariesController],
-  providers: [BinariesService, BinaryResourcesService, BinariesInstallerService],
-  exports: [BinariesService, BinaryResourcesService]
+  providers: [BinariesService, BinaryResourcesService, BinariesInstallerService, OfflinePackageService],
+  exports: [BinariesService, BinaryResourcesService, OfflinePackageService]
 })
 export class BinariesModule {}